Configure the Jenkins URL with Groovy
This change adds an environment variable to the compose file that sets
the URL of the Jenkins instance upon boot. This is done via the script
added to init.groovy.d.

This change adds a script to bring down and up the containers cleanly.
This is done via the script start.sh.